fix uniswap helpers linking
