+++ governance/helpers/uniswap.js
@@ -1,37 +1,17 @@
 const bn = require('bignumber.js')
 const { ethers } = require('hardhat')
-const { Pool, Tick, TickListDataProvider } = require('@uniswap/v3-sdk/')
-const {
-  Token,
-  CurrencyAmount,
-  TradeType,
-  Percent,
-} = require('@uniswap/sdk-core')
-const {
-  AlphaRouter,
-  SwapType,
-  nativeOnChain,
-  MAX_UINT160,
-} = require('@uniswap/smart-order-router')
-const JSBI = require('jsbi')
-const { PERMIT2_ADDRESS } = require('@uniswap/universal-router-sdk')
-const { AllowanceTransfer } = require('@uniswap/permit2-sdk')
+const { Token, Pool, Tick, TickListDataProvider } = require('@uniswap/v3-sdk')
 const {
   ADDRESS_ZERO,
   MAX_UINT,
   BASIS_POINTS,
   WETH,
-  USDC,
-  DAI,
   UDT,
-  WBTC,
   UNISWAP_FACTORY_ADDRESS,
-  V3_SWAP_ROUTER_ADDRESS,
   POSITION_MANAGER_ADDRESS,
-  CHAIN_ID,
   getERC20Contract,
+  getTokenInfo,
 } = require('@unlock-protocol/hardhat-helpers')
-const ERC20_ABI = require('@unlock-protocol/hardhat-helpers/dist/ABIs/erc20.json')
 
 // default fee
 const FEE = 500
@@ -89,18 +69,6 @@ async function getPoolState(poolContract) {
   }
 
   return PoolState
-}
-
-const getTokenInfo = async (tokenAddress) => {
-  const token0 = await ethers.getContractAt(ERC20_ABI, tokenAddress)
-  const [decimals, symbol] = await Promise.all([
-    await token0.decimals(),
-    await token0.symbol(),
-  ])
-  return {
-    decimals,
-    symbol,
-  }
 }
 
 const getTokens = async (pool) => {
@@ -323,132 +291,11 @@ const deployUniswapV3Oracle = async function () {
   return oracle
 }
 
-const currencyAmountToBigNumber = (amount) => {
-  const { decimals } = amount.currency
-  const fixed = ethers.FixedNumber.from(amount.toExact())
-  const tokenScale = ethers.FixedNumber.from(
-    ethers.BigNumber.from(10).pow(decimals)
-  )
-  return ethers.BigNumber.from(
-    // have to remove trailing .0 "manually" :/
-    fixed.mulUnsafe(tokenScale).floor().toString().split('.')[0]
-  )
-}
-
-/**
- * PERMIT2 helpers
- * */
-const makePermit = (
-  tokenAddress,
-  amount = ethers.constants.MaxUint256.toString(),
-  deadline = Math.floor(new Date().getTime() / 1000 + 100000),
-  nonce = '0'
-) => {
-  return {
-    details: {
-      token: tokenAddress,
-      amount,
-      expiration: deadline.toString(),
-      nonce,
-    },
-    spender: V3_SWAP_ROUTER_ADDRESS,
-    sigDeadline: deadline.toString(),
-  }
-}
-
-async function generatePermitSignature(permit, signer, chainId = CHAIN_ID) {
-  const { domain, types, values } = AllowanceTransfer.getPermitData(
-    permit,
-    PERMIT2_ADDRESS,
-    chainId
-  )
-  return await signer._signTypedData(domain, types, values)
-}
-
-/**
- * UNISWAP ROUTER
- * */
-async function getUniswapRoute({
-  tokenIn,
-  tokenOut,
-  amoutOut = ethers.utils.parseUnits('10', tokenOut.decimals),
-  recipient,
-  slippageTolerance = new Percent(10, 100),
-  deadline = Math.floor(new Date().getTime() / 1000 + 100000),
-  permitOptions: { usePermit2Sig = false, inputTokenPermit } = {},
-  chainId = CHAIN_ID,
-}) {
-  // init router
-  const router = new AlphaRouter({
-    chainId,
-    provider: ethers.provider,
-  })
-
-  // parse router args
-  const outputAmount = CurrencyAmount.fromRawAmount(
-    tokenOut,
-    JSBI.BigInt(amoutOut)
-  )
-  const routeArgs = [
-    outputAmount,
-    tokenIn,
-    TradeType.EXACT_OUTPUT,
-    {
-      type: SwapType.UNIVERSAL_ROUTER,
-      recipient,
-      slippageTolerance,
-      deadline,
-    },
-  ]
-
-  // add sig if necessary
-  if (usePermit2Sig) routeArgs.inputTokenPermit = inputTokenPermit
-
-  // call router
-  const { methodParameters, quote, quoteGasAdjusted, estimatedGasUsedUSD } =
-    await router.route(...routeArgs)
-
-  // parse quote as BigNumber
-  const amountInMax = currencyAmountToBigNumber(quote)
-
-  // log some prices
-  console.log(`Quote Exact: ${quote.toExact()}`)
-  console.log(`Quote toFixed: ${quote.toFixed(2)}`)
-  console.log(`Gas Adjusted Quote: ${quoteGasAdjusted.toFixed(2)}`)
-  console.log(`Gas Used USD: ${estimatedGasUsedUSD.toFixed(6)}`)
-  console.log(`AmountInMax: ${amountInMax.toString()}`)
-
-  const { calldata: swapCalldata, value, to: swapRouter } = methodParameters
-
-  return {
-    swapCalldata,
-    value,
-    amountInMax,
-    swapRouter,
-  }
-}
-
-const getUniswapTokens = (chainId = 1) => ({
-  native: nativeOnChain(chainId),
-  dai: new Token(chainId, DAI, 18, 'DAI'),
-  weth: new Token(chainId, WETH, 18, 'WETH'),
-  usdc: new Token(chainId, USDC, 6, 'USDC'),
-  udt: new Token(chainId, UDT, 18, 'UDT'),
-  wBtc: new Token(chainId, WBTC, 18, 'wBTC'),
-})
-
 module.exports = {
   addLiquidity,
   createUniswapV3Pool: createPool,
   deployUniswapV3Oracle,
-  currencyAmountToBigNumber,
   getTokenInfo,
   getPoolState,
   getPoolImmutables,
-  getUniswapRoute,
-  getUniswapTokens,
-  PERMIT2_ADDRESS,
-  MAX_UINT160,
-  makePermit,
-  generatePermitSignature,
 }
