+++ governance/package.json
@@ -9,9 +9,6 @@
     "@safe-global/safe-core-sdk": "3.3.5",
     "@safe-global/safe-ethers-lib": "1.9.3",
     "@safe-global/safe-service-client": "2.0.3",
-    "@uniswap/smart-order-router": "3.17.2",
-    "@uniswap/universal-router-sdk": "2.0.2",
-    "@uniswap/v3-sdk": "3.10.0",
     "@unlock-protocol/contracts": "workspace:./packages/contracts",
     "@unlock-protocol/eslint-config": "workspace:./packages/eslint-config",
     "@unlock-protocol/hardhat-helpers": "workspace:^",
