+++ packages/hardhat-helpers/package.json
@@ -16,6 +16,10 @@
     "@nomiclabs/hardhat-ethers": "2.2.3",
     "@nomiclabs/hardhat-waffle": "2.0.6",
     "@openzeppelin/hardhat-upgrades": "1.28.0",
+    "@uniswap/permit2-sdk": "1.2.0",
+    "@uniswap/sdk-core": "4.0.9",
+    "@uniswap/smart-order-router": "3.19.0",
+    "@uniswap/v3-sdk": "3.10.0",
     "@unlock-protocol/contracts": "workspace:^",
     "@unlock-protocol/networks": "workspace:^",
     "@unlock-protocol/tsconfig": "workspace:^",
