+++ packages/hardhat-helpers/src/constants.ts
@@ -2,6 +2,7 @@ export default {
   ADDRESS_ZERO: '0x0000000000000000000000000000000000000000',
   MAX_UINT:
     '115792089237316195423570985008687907853269984665640564039457584007913129639935',
+  MAX_UINT160: '0xffffffffffffffffffffffffffffffffffffffff',
   MAX_GAS: 6700000,
   BASIS_POINTS: 1000000,
 }
